tr1/output: move good functions to its own module
