+++ src/tr1/meson.build
@@ -244,12 +244,15 @@ sources = [
   'game/option/option_passport.c',
   'game/option/option_sound.c',
   'game/output.c',
+  'game/output/draw_misc.c',
+  'game/output/func.c',
   'game/output/meshes/common.c',
   'game/output/meshes/dynamic.c',
   'game/output/meshes/objects.c',
   'game/output/meshes/rooms.c',
   'game/output/shader.c',
   'game/output/sprites.c',
+  'game/output/state.c',
   'game/output/textures.c',
   'game/overlay.c',
   'game/requester.c',
